refactor(tibuild): refactor product mapping and service constants
Extract them in to file `map.go.
